requirements/base.txt
django==1.11.*
django-countries>=3.1.1
pyuca>=1.1.2
dj-database-url>=0.4.2
whitenoise==3.3.*
django-libsass==0.7
raven==6.7.0
django-solo==1.1.3
Pillow==5.1.0
